✨ Feat: Directory Moved (✨ Routing + Code Performance Improvement #12)
